Fixed g++-10 -Wall warnings (#431)
* Fixed 'multi-character character constant' warning
* Fixed g++-10 -Wsign-compare warnings
* Fixed g++-10 -Wreorder warnings.
* Fixed g++-10 -Wswitch (enumeration value not handled in switch) warnings
* Fixed g++-10 -Wunused-but-set-variable warnings.
* Fixed g++-10 -Wunknown-pragmas warnings.
* Remove #pragma warning for SOIL and GLI/GLM
Will produce warnings on MSVC, have to think of a different solution to silence them on VS.

+++ src/components/UndoHistory.h
@@ -26,7 +26,7 @@ public:
 
 	bool CanUndo() const { return curIndex != UH_NONE; }
 
-	bool CanRedo() const { return !states.empty() && curIndex + 1 < static_cast<int>(states.size()); }
+	bool CanRedo() const { return !states.empty() && curIndex + 1 < states.size(); }
 
 	UndoStateProject* GetCurState() const {
 		if (curIndex == UH_NONE)
@@ -41,7 +41,7 @@ public:
 	}
 
 	UndoStateProject* GetNextState() const {
-		if (curIndex + 1 >= static_cast<int>(states.size()))
+		if (curIndex + 1 >= states.size())
 			return nullptr;
 		return states[curIndex + 1].get();
 	}
